% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg-tgc.R
\name{as_tgc}
\alias{as_tgc}
\alias{as_tgc.default}
\alias{as_tgc.tgc}
\alias{as_tgc.tg_geometrycollection}
\alias{as_tgc.tg}
\alias{as_tgc.data.frame}
\title{Coerce to a TG geometry collection}
\usage{
as_tgc(x, ...)

\method{as_tgc}{default}(x, ...)

\method{as_tgc}{tgc}(x, ...)

\method{as_tgc}{tg_geometrycollection}(x, ...)

\method{as_tgc}{tg}(x, ...)

\method{as_tgc}{data.frame}(x, geom_col, ...)
}
\description{
Coerce to a TG geometry collection
}
