% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg-tgc.R
\name{str.tgc}
\alias{str.tgc}
\title{pulled from geos::str.geos_geometry}
\usage{
\method{str}{tgc}(object, ..., indent.str = "", width = getOption("width"))
}
\description{
pulled from geos::str.geos_geometry
}
