% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg-io.R
\name{tg_read_wkb}
\alias{tg_read_wkb}
\alias{tg_read_wkt}
\alias{tg_read_geojson}
\title{Read WKB into TG geometry}
\usage{
tg_read_wkb(x, ..., crs = NULL)

tg_read_wkt(x, ..., crs = NULL)

tg_read_geojson(x, ..., crs = NULL)
}
\arguments{
\item{x}{GeoJSON string vector}
}
\description{
Read WKB into TG geometry

Read WKT into TG geometry

Read a GeoJSON string into TG geometry
}
