% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg-io.R
\name{as_tg}
\alias{as_tg}
\alias{as_tg.default}
\alias{as_tg.tg}
\alias{as_tg.character}
\alias{as_tg.blob}
\alias{as_tg.WKB}
\alias{as_tg.raw}
\title{Coerce to a TG geometry}
\usage{
as_tg(x, ...)

\method{as_tg}{default}(x, ...)

\method{as_tg}{tg}(x, ...)

\method{as_tg}{character}(x, ..., crs = NULL)

\method{as_tg}{blob}(x, ..., crs = NULL)

\method{as_tg}{WKB}(x, ..., crs = NULL)

\method{as_tg}{raw}(x, ..., crs = NULL)
}
\arguments{
\item{x}{An object interpretable of type character or raw.}

\item{...}{unused}

\item{crs}{An ESPG ID}
}
\value{
A TG geometry vector
}
\description{
Coerce to a TG geometry
}
\examples{
as_tg_geometry("POINT(1.0 2.0)")

gj <- '{"type":"Feature","geometry":{"type":"Point","coordinates":[125.6,10.1]},"properties":{"name":"Dinagat Islands"}}'
as_tg_geometry(gj)

}
\seealso{
\code{\link[=tg_read_geojson]{tg_read_geojson()}} for GeoJSON,
\code{\link[=tg_read_wkt]{tg_read_wkt()}} for WKT, and \code{\link[=tg_read_wkb]{tg_read_wkb()}} for WKB
}
